% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ambientais_loop.r
\name{Ambientais_loop}
\alias{Ambientais_loop}
\title{esta função, com base nas entradas modo_radiacao, TipoAlbedo, TipoTA, TipoVV, retorna
os valores de radiacao, albedo, temperatura ambiente e velocidade do vento}
\usage{
Ambientais_loop(modo_radiacao, G.cnb, G.cd, theta.z, G.o, DadosAmbientais,
  HaSol, AlbedoFixo, TipoAlbedo, VVFixo, TipoVV, TAFixo, TipoTA, Ibn, Id,
  T.Sintetico, tempos, DadosMensais)
}
\description{
esta função, com base nas entradas modo_radiacao, TipoAlbedo, TipoTA, TipoVV, retorna
os valores de radiacao, albedo, temperatura ambiente e velocidade do vento
}
