% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bruta.r
\name{fbruta}
\alias{fbruta}
\title{Brute Force}
\usage{
fbruta(g, a, r, p, m, perdasTot, passo, TempRef, n_inv, pot_inv, Pdc0, IrrRef,
  n_non, n_ref, PE)
}
\description{
This function makes a sweep trough gama and Beta
and stores value of energy and value generated. No output yet.
Until this point, this function is beeing used mannualy.
}
