% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCusto.R
\name{fCusto}
\alias{fCusto}
\title{The difference between fCusto and fCusto2, is that fCusto2 returns, also, the value of generated energy
 @export}
\usage{
fCusto(Beta_gama, g, a, r, p, m, perdasTot, passo, TempRef, n_inv, pot_inv,
  Pdc0, IrrRef, n_non, n_ref, PE)
}
\description{
The difference between fCusto and fCusto2, is that fCusto2 returns, also, the value of generated energy
 @export
}
