% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCusto2.R
\name{fCusto2}
\alias{fCusto2}
\title{The difference between fCusto and fCusto2, is that fCusto2 returns, also, the value of generated energy}
\usage{
fCusto2(Beta_gama, g, a, r, p, m, perdasTot, passo, TempRef, n_inv, pot_inv,
  Pdc0, IrrRef, n_non, n_ref, PE)
}
\description{
The difference between fCusto and fCusto2, is that fCusto2 returns, also, the value of generated energy
}
