% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTracker2.R, R/GeoTracker3.R
\name{geo.trackerNREL}
\alias{geo.trackerNREL}
\alias{geo.trackerNREL}
\title{Função geométrico com tracker.}
\usage{
geo.trackerNREL(lat, long, long.st, Beta, gama, passo, Hora_do_passo,
  tempo.inicial, tempo.final, Rastreamento)

geo.trackerNREL(lat, long, long.st, Beta, gama, passo, Hora_do_passo,
  tempo.inicial, tempo.final, Rastreamento)
}
\arguments{
\item{lat}{Latidude}

\item{long}{Longitude}

\item{long.st}{Longitudo da hora padrao}

\item{Beta}{Declinacao do painel}

\item{gama}{Orientacao}

\item{passo}{tamanho do passo em minutos. default para 60}

\item{tempo.inicial, tempo.final}{strings no formato "dd/mm/aaaa hh:mm" para definir o tempo final e inicial. Default para "01/01/2015 00:00"}

\item{hora_do_passo}{informa o momento em que as medicoes para cada passo sao levadas em conta. Aceita "Mais_tarde", "exata". Default para "Mais_tarde. Qualquer outro input considera como mais cedo}

\item{lat}{Latidude}

\item{long}{Longitude}

\item{long.st}{Longitudo da hora padrao}

\item{Beta}{Declinacao do painel}

\item{gama}{Orientacao}

\item{passo}{tamanho do passo em minutos. default para 60}

\item{hora_do_passo}{informa o momento em que as medicoes para cada passo sao levadas em conta. Aceita "Mais_tarde", "exata". Default para "Mais_tarde. Qualquer outro input considera como mais cedo}

\item{tempo.inicial, tempo.final}{strings no formato "dd/mm/aaaa hh:mm" para definir o tempo final e inicial. Default para "01/01/2015 00:00"}
}
\description{
Esta funcao recebe como entradas informacoes da posicao do posicionamento do painel
e de tempo de simulacao e
retorna informacoes sobre os angulos solares por hora, como angulos horarios, declinacao, angulo azimutal. 
retorna tambem o vetor de tempo.

Esta funcao recebe como entradas informacoes da posicao do posicionamento do painel
e de tempo de simulacao e
retorna informacoes sobre os angulos solares por hora, como angulos horarios, declinacao, angulo azimutal. 
retorna tambem o vetor de tempo.
}
\details{
Esta funcao e o pilar do programa, varias de suas saidas sao utilizadas por quase todas as outras funcoes

Esta funcao e o pilar do programa, varias de suas saidas sao utilizadas por quase todas as outras funcoes
}
