% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rastreado.R
\name{Geracao_de_Energia_Rastreada}
\alias{Geracao_de_Energia_Rastreada}
\title{Geracao_de_Energia_Rastreada}
\usage{
Geracao_de_Energia_Rastreada(long, long.st, lat, A, DadosAmbientais, Pdc0, Beta,
  gama, TipoDePainel, pot_inv, n_inv = 1L, n_non, perdasTot, G.sc = 1367,
  TempRef = 25, IrrRef = 1000, n_ref = 0.9637,
  tempo.inicial = "01/01/2015 00:00", tempo.final = "31/12/2015 23:00",
  passo = 60, Hora_do_passo = "Mais_tarde", AlbedoFixo = 0.2,
  TipoAlbedo = "fixo", VVFixo = 1.7, TipoVV = "fixo", TaFixo = 25,
  TipoTA = "fixo", modo_radiacao = "ceu_limpo", ModeloSintese = "Liu",
  ModeloComposicao = "Collares", TipoDeClima, TipoModulo_Montagem,
  DadosMensais, TemperaturaConstante, Rastreamento)
}
\arguments{
\item{long}{longitude in degrees.
Negative values for south. Acceped values in between -180 and 180}

\item{long.st}{standard longitude of the timezone.}

\item{lat}{latitde in degrees.
Negative values for west. Accepted values in betwwen -60 and 60}

\item{A}{Altitude. Used to calculate clear-sky radiation}

\item{DadosAmbientais}{Data frame containing the hourly
environmental input data, if any is used}

\item{Pdc0}{nominal dc power of the system}

\item{Beta}{Declination angle in degrees.
Accepted values in betwwen 0 and 90}

\item{gama}{Orientation angle. N=0, E=90, S=180, W=270}

\item{TipoDePainel}{Type of pannel. Accepted inputs:
"Cristalino Premium"; "Cristalino Padrao";"Filme fino"
Other values will be considered as "outro". Used in
CaracteristicPainel(TipoDePainel)}

\item{pot_inv}{Inverter nominal power}

\item{n_inv}{number of inverters}

\item{perdasTot}{Total external losses.}

\item{G.sc}{Solar Constant}

\item{TempRef}{Reference temperature of modules}

\item{IrrRef}{Reference irradiation on the module}

\item{n_ref}{efficiency reference}

\item{tempo.inicial}{Initial time of simulation}

\item{tempo.final}{Final time of simulation}

\item{passo}{simulation step in minutes}

\item{Hora_do_passo}{when to consider the value for the
simultion step. Available options:
"Mais_tarde", "Mais_cedo", "Exata"}

\item{AlbedoFixo}{Value for albedo if it is set as fixed}

\item{TipoAlbedo}{Type of albedo.
Available options:
"fixo", "mensal", "horario"}

\item{VVFixo}{Value for wind-speed if it is set as fixed}

\item{TipoVV}{Type of wind-speed.
Available options:
"fixo", "mensal", "horario"}

\item{modo_radiacao}{Type of radiation data used.
Accepted values:
"ceu_limpo", "Kt_mes", "H_med", "horario"}

\item{ModeloSintese}{"KTmedio","LiuJordanModificado","Liu","Knight","Whillier","WhillierModificado"}

\item{ModeloComposicao}{"Orgill","Erbs","Collares","CollaresD"}

\item{TipoModulo_Montagem}{type of mount of system. used in thermal moddeling
"Polimero/filme-fino/aco_Estrutura","Vidro/celula/polimero_Estrutura",
"Vidro/celula/polimero_Telhado","Vidro/celula/vidro_Estrutura","Vidro/celula/vidro_Telhado"}

\item{DadosMensais}{Monthly dada input dataframe, if used}

\item{TemperaturaConstante}{1 if there is no sthochastic component on sinthetic temperature}

\item{Rastreamento}{type of tracking.
"fixo","1eixo","2eixos","azimutal"}

\item{n_nom}{Nominal Efficiency}

\item{TAFixo}{Value for Environment Temperature if it is set as fixed}

\item{TipoAlbedo}{Type of Environment Temperature.
Available options:
"fixo", "mensal", "horario"}

\item{TipoDeCilma}{clime type, used to create clear sky radiations.
"Lat.intermed./Inverno","Lat.intermed./Verao","Tropical","Verao Subartico"}
}
\description{
This function executes the whole simulation of the sun generation
It can take many forms of inputs
}
\author{
André Costa
}
