% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodo_gradiente.R
\name{grad}
\alias{grad}
\title{Gradient method}
\usage{
grad(Beta_gama0, g, a, r, p, m, perdasTot, passo, TempRef, n_inv, pot_inv, Pdc0,
  IrrRef, n_non, n_ref, PE)
}
\description{
This function uses the gradient method to find an opmal x =c(Beta, gama), minimizes fcusto
}
