% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mes2dia.R
\name{Mes2Dia}
\alias{Mes2Dia}
\title{Funcao que pega dados mensais e retorna valores de Kt medios diarios. Baseado em Liu e Jordan.
A entrada DadosMensais deve conter as colunas:
n.Dias, com o numero de dias do referido mes,
Gama (que é funcao do Kt.min, Kt.max e Kt.mes)
Kt.min
Kt.max
Kt.mes}
\usage{
Mes2Dia(DadosMensais)
}
\description{
A Função retorna um vetor com tamanho 365(ou diferente, se a soma de n.dias for outra)
Este vetor contem os valores gerados de Kt diario
}
\details{
Observação: para fazer a permutação dos dias em meses com menos de 31 dias, o dia 31 foi desconsiderado
Observação: Pode ser interessante criar
}
