% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemperaturaSintetico.R
\name{TemperaturaSintetico}
\alias{TemperaturaSintetico}
\title{Temperatura Sintetico}
\usage{
TemperaturaSintetico(tempos, DadosMensais, tempos.solar, TemperaturaConstante)
}
\description{
Essa funcao recebe informacoes mensais (sao necessarias Kt.mes e TA) e o conjunto de tempos, gerando temperaturas com intervalo de uma hora
}
