% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrabalhaDados.R
\name{TrabalhaDados}
\alias{TrabalhaDados}
\title{essa funcao transforma H.med em Kt.mes. Tambem cria Kt.min, Kt.max, Gama, Csi, caso nao estejam na entrada}
\usage{
TrabalhaDados(DadosMensais, lat, G.sc)
}
\description{
essa funcao transforma H.med em Kt.mes. Tambem cria Kt.min, Kt.max, Gama, Csi, caso nao estejam na entrada
}
