% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrabalhaDados2.R
\name{TrabalhaDados2}
\alias{TrabalhaDados2}
\title{essa funcao transforma H.med em Kt.mes. Tambem cria Kt.min, Kt.max, Gama, Csi, caso nao estejam na entrada}
\usage{
TrabalhaDados2(DadosMensais, lat, G.sc)
}
\description{
essa funcao transforma H.med em Kt.mes. Tambem cria Kt.min, Kt.max, Gama, Csi, caso nao estejam na entrada
}
