% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransmitanciaAtmosferica.r
\name{TransmitanciaAtmosferica}
\alias{TransmitanciaAtmosferica}
\title{This funcion gives as outputs the parameters r.0 r.1 and r.k of atmospheric transmitnace considering the input TipoDeClima}
\usage{
TransmitanciaAtmosferica(TipoDeClima)
}
\description{
This funcion gives as outputs the parameters r.0 r.1 and r.k of atmospheric transmitnace considering the input TipoDeClima
}
